lab7 - hashing
Write a program using the modulo-division method and linear probing to store a list of inventory
parts and their quantities sold in the past month as shown below in an array with the maximum size
as 20.

Perform the following operations
a. Search for an inventory item and report its quantity sold
b. Determine how many collisions occurred

Part Number  Quantity
112 12
130 30
156 56
173 17
197 19
150 50
166 66
113 13
123 12
143 14
167 16
189 18
193 19
117 11
176 76